feat(#435): exact MTP verify is now OPT-IN via --exact-verify, not the default
Inverts the #435 PR1 default. The verify pass now runs the legacy WY-chunkwise
/ fused BF16-conv arms by default (the pre-#435 numerics), and the
sequential-decode-exact chain is selected by a new --exact-verify flag.
--verify-wy is REMOVED, not deprecated: it never shipped in a release, and
with the polarity inverted it would be a second name for the default state.

Why invert: exact verify was measured on GPU at the widths the ladder
actually verifies at — +116%/+63%/+69% on the GDN phase at n=8/16/32
(~+35%/+22%/+36% per decode step). Every surveyed production engine ships
exactness as an opt-in (vLLM VLLM_BATCH_INVARIANT, SGLang
--enable-deterministic-inference at ~34% avg slowdown, TensorRT-LLM/TGI
none); our cost sits inside that range. Decided by the swarm review; width
gating was explicitly rejected (no precedent, recreates batch-size-dependent
output changes).

Honesty: the docs now state plainly that with default settings #435's
divergence REMAINS — spec-on output is NOT bitwise-equal to spec-off at
temp 0 — and name --exact-verify with its measured cost. The dispatch-site
comments, gdn_flags docs, serve help text, engage-log lines and the
microtest header all say opt-in, none imply #435 is fixed by default.

Tests: verify_exact_is_the_default / verify_wy_and_h_f16_each_disable_exact
encoded the OLD default and are replaced (not silently flipped) by
legacy_wy_verify_is_the_default, exact_verify_flag_selects_the_exact_chain,
env_fallback_never_enables_exact_verify and
h_f16_forces_non_exact_even_when_requested — the expectation changed because
the default changed, by decision. New validator rule: an explicit
--exact-verify beside --ssm-h-dtype f16 is a hard error (the exact arm's
FP32 kernels must never read an FP16 pool; silently dropping the opt-in
would ship the divergence the operator just opted out of); the pure-function
h_f16 guard stays as defense in depth. All five targeted mutations
(polarity, h_f16 guard, env default, bare-flag value, validator rule) go red
on the intended assertion; suite is 3947/0 with fmt+clippy clean.

validate.rs tests moved to a sibling validate_tests.rs via the
helpers_tests.rs #[path] idiom to stay under the 500-line cap;
x435_ssm_cost.rs (584 LoC, landed with the #435 stack) is allow-listed per
the dev-example precedent.
