+++ crates/spark-model/src/layers/qwen3_ssm/trait_decode_batched_conv_gdn_exact.rs
@@ -1,13 +1,21 @@
 // SPDX-License-Identifier: AGPL-3.0-only
 
-//! Exact MTP-verify conv+GDN arm (issue #435 route (a), PR1).
+//! Exact MTP-verify conv+GDN arm (issue #435 route (a)) — OPT-IN via
+//! `--exact-verify`.
 //!
-//! At temp 0, spec-on output must equal spec-off output. The WY/fused verify
-//! arms broke that in two ways: the verify conv ran the BF16-output kernel
-//! where sequential decode runs the FP32 one (h-state relL2 ~8.6e-4 after one
-//! K=4 window — the dominant term), and the WY chunkwise reordering differs
-//! from the recurrent update by ~3.4e-8 — four orders smaller but nonzero,
-//! and an argmax flip only needs a per-logit error above a thin top-2 margin.
+//! By default this arm does NOT run: the verify pass runs the WY/fused
+//! chunkwise arms, so spec-on output is NOT bitwise-equal to spec-off at
+//! temp 0 — #435's divergence remains with default settings, and
+//! `--exact-verify` is the flag that removes it (measured decode-step cost
+//! ~+35% at n=8/K=4, ~+22% at n=16/K=2, ~+36% at n=32/K=2; every surveyed
+//! production engine ships exactness as the same kind of opt-in).
+//!
+//! The WY/fused verify arms diverge in two ways: the verify conv runs the
+//! BF16-output kernel where sequential decode runs the FP32 one (h-state
+//! relL2 ~8.6e-4 after one K=4 window — the dominant term), and the WY
+//! chunkwise reordering differs from the recurrent update by ~3.4e-8 — four
+//! orders smaller but nonzero, and an argmax flip only needs a per-logit
+//! error above a thin top-2 margin.
 //!
 //! This arm runs, per verify token, EXACTLY the kernel chain `ssm_forward`
 //! (the single-token decode reference) runs — same handles, same launch
@@ -30,10 +38,10 @@
 //! phase-8 norm when `verify_exact_enabled()` — both sites read that one
 //! predicate.
 //!
-//! Kill switch: `--verify-wy` restores the WY/fused arms for A/B. Default is
-//! exact, because this is a correctness fix. Every dispatch decision below is
-//! a pure function of (process-static flags, kernel handles, the slot's fixed
-//! intermediate addresses), so it is CUDA-graph-stable.
+//! Kill switch: omit `--exact-verify` (the default) to run the WY/fused arms.
+//! Every dispatch decision below is a pure function of (process-static flags,
+//! kernel handles, the slot's fixed intermediate addresses), so it is
+//! CUDA-graph-stable.
 
 use anyhow::Result;
 use spark_runtime::gpu::DevicePtr;
@@ -159,10 +167,10 @@ impl Qwen3SsmLayer {
         static LOGGED: std::sync::Once = std::sync::Once::new();
         LOGGED.call_once(|| {
             tracing::info!(
-                "EXACT MTP verify ENGAGED (#435 route (a)): per-token sequential-decode \
-                 kernel chain (f32_conv={use_f32_conv}, fused_gdn_norm={fused_gdn_norm}, \
-                 snap_twin={snap}, fused_f32_conv={fused_conv}); WY arms restorable \
-                 with --verify-wy"
+                "EXACT MTP verify ENGAGED (#435 route (a), opt-in --exact-verify): \
+                 per-token sequential-decode kernel chain (f32_conv={use_f32_conv}, \
+                 fused_gdn_norm={fused_gdn_norm}, snap_twin={snap}, \
+                 fused_f32_conv={fused_conv}); omit the flag for the default WY arms"
             );
         });
