+++ crates/spark-model/src/layers/qwen3_ssm/trait_decode_batched_conv_gdn_multi.rs
@@ -97,11 +97,12 @@ impl Qwen3SsmLayer {
     ) -> Result<bool> {
         let n = states.len();
         let kk = args.num_tokens;
-        // ── Issue #435 route (a), DEFAULT: exact batched verify (strided
-        // sequential-decode chain at batch=n, bitwise-equal to spec-off
-        // decode). Ok(false) → the caller's per-sequence loop, which under
-        // the same predicate runs the per-token exact arm — identical bits.
-        // The WY fast path below survives behind `--verify-wy` for A/B.
+        // ── Issue #435 route (a), OPT-IN via `--exact-verify`: exact batched
+        // verify (strided sequential-decode chain at batch=n, bitwise-equal
+        // to spec-off decode). Ok(false) → the caller's per-sequence loop,
+        // which under the same predicate runs the per-token exact arm —
+        // identical bits. The WY fast path below is the DEFAULT; without the
+        // flag, spec-on output is NOT bitwise-equal to spec-off (#435).
         if super::verify_exact_enabled() {
             return self.decode_batched_conv_gdn_multi_exact(states, ctx, args);
         }
