+++ crates/spark-server/src/cli/validate.rs
@@ -92,6 +92,20 @@ pub fn validate_serve_args(args: &ServeArgs) -> Result<(), String> {
             "add --gdn-fused-norm, or use --ssm-h-dtype f32",
         ));
     }
+    // #435: the exact-verify arm's kernels are FP32 readers, so an FP16
+    // h-state pool disables it (`GdnFlags::verify_exact_active`). Honouring
+    // `--ssm-h-dtype f16` by SILENTLY ignoring an explicit `--exact-verify`
+    // would ship the very divergence the operator just opted out of — the
+    // exact class of combination this validator exists to refuse.
+    if args.exact_verify == Some(true) && args.ssm_h_dtype.as_deref() == Some("f16") {
+        v.push(Violation::new(
+            "--exact-verify together with --ssm-h-dtype f16",
+            "the exact MTP-verify chain (issue #435) runs FP32-reader kernels and must \
+             never read the FP16 h-state pool, so with f16 the exact request would be \
+             silently dropped and spec-on output would NOT equal spec-off",
+            "drop --ssm-h-dtype f16 (f32 is the default), or drop --exact-verify",
+        ));
+    }
     check_enum(
         &mut v,
         "--mtp-quantization",
@@ -283,214 +297,5 @@ fn format_violations(v: &[Violation]) -> String {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use clap::Parser;
-
-    /// Parse a `spark serve ...` command line into `ServeArgs` for testing.
-    fn parse(extra: &[&str]) -> ServeArgs {
-        let mut argv = vec!["spark", "serve", "dummy/model", "--model-name", "dummy"];
-        argv.extend_from_slice(extra);
-        match super::super::Cli::parse_from(argv).command {
-            super::super::Command::Serve(a) => a,
-            super::super::Command::Benchmark(_) => unreachable!("this test parses a serve command"),
-        }
-    }
-
-    #[test]
-    fn defaults_are_valid() {
-        assert!(validate_serve_args(&parse(&[])).is_ok());
-    }
-
-    #[test]
-    fn fp8_calibration_requires_fp8_kv() {
-        let err = validate_serve_args(&parse(&[
-            "--kv-cache-dtype",
-            "bf16",
-            "--fp8-kv-calibration-tokens",
-            "256",
-        ]))
-        .unwrap_err();
-        assert!(err.contains("--fp8-kv-calibration-tokens"));
-        assert!(err.contains("fix:"));
-        // The same flags with an fp8 cache are fine.
-        assert!(
-            validate_serve_args(&parse(&[
-                "--kv-cache-dtype",
-                "fp8",
-                "--fp8-kv-calibration-tokens",
-                "256",
-            ]))
-            .is_ok()
-        );
-    }
-
-    #[test]
-    fn an_absent_lever_flag_parses_as_unspecified() {
-        // `None` is not the same as the default VALUE, and the difference is
-        // load-bearing: publishing a default seals the cell these two flags
-        // write to, which turned `ATLAS_SSM_TAIL_MIDCHUNK=0` and
-        // `ATLAS_MTP_GATE_FORCE=1` into documented, echoed, silent no-ops under
-        // `spark serve`. Absent has to stay absent all the way to
-        // `publish_kernel_flags` for the fallback to be reachable.
-        let a = parse(&[]);
-        assert!(a.ssm_tail_midchunk.is_none(), "ATLAS_SSM_TAIL_MIDCHUNK");
-        assert!(a.mtp_gate.is_none(), "ATLAS_MTP_GATE_FORCE");
-        assert!(a.ssm_h_dtype.is_none(), "ATLAS_SSM_H_FP16");
-        assert!(a.gdn_fused_norm.is_none(), "ATLAS_GDN_FUSED_NORM");
-        assert!(
-            a.ssm_batched_recurrent.is_none(),
-            "ATLAS_SSM_BATCHED_RECURRENT"
-        );
-        // #435: absent must stay absent so publish_kernel_flags does not seal
-        // the GDN cell; the resolved default (exact verify) is asserted in
-        // gdn_flags' own tests.
-        assert!(a.verify_wy.is_none(), "--verify-wy");
-
-        let a = parse(&["--ssm-tail-midchunk", "false", "--mtp-gate", "force"]);
-        assert_eq!(a.ssm_tail_midchunk, Some(false), "given, it still wins");
-        assert_eq!(a.mtp_gate.as_deref(), Some("force"));
-    }
-
-    #[test]
-    fn the_bare_gdn_switches_still_mean_on() {
-        // `Option<bool>` must not turn a presence switch into one that DEMANDS
-        // a value: every recipe and frozen command line writes them bare, and
-        // silently requiring `--gdn-fused-norm true` would break all of them.
-        let a = parse(&["--gdn-fused-norm", "--ssm-batched-recurrent"]);
-        assert_eq!(a.gdn_fused_norm, Some(true));
-        assert_eq!(a.ssm_batched_recurrent, Some(true));
-        // And an explicit off is now expressible, which it was not before.
-        let a = parse(&["--gdn-fused-norm", "false"]);
-        assert_eq!(a.gdn_fused_norm, Some(false));
-        // #435 A/B kill switch follows the same convention: bare means on
-        // (restore the WY arms), explicit false is expressible.
-        let a = parse(&["--verify-wy"]);
-        assert_eq!(a.verify_wy, Some(true));
-        let a = parse(&["--verify-wy", "false"]);
-        assert_eq!(a.verify_wy, Some(false));
-    }
-
-    #[test]
-    fn f16_h_state_still_needs_the_fused_norm_arm() {
-        // Absent counts as off: one GDN flag publishes all three, so
-        // `--ssm-h-dtype f16` alone reaches the FP32-only kernel with an FP16
-        // pool — fluent garbage, not a fault.
-        assert!(validate_serve_args(&parse(&["--ssm-h-dtype", "f16"])).is_err());
-        assert!(
-            validate_serve_args(&parse(&["--ssm-h-dtype", "f16", "--gdn-fused-norm"])).is_ok(),
-            "the supported pairing"
-        );
-        assert!(
-            validate_serve_args(&parse(&[
-                "--ssm-h-dtype",
-                "f16",
-                "--gdn-fused-norm",
-                "false"
-            ]))
-            .is_err(),
-            "and an explicit off is refused just as an absent one is"
-        );
-    }
-
-    #[test]
-    fn a_mistyped_mtp_gate_is_still_caught() {
-        // Making the flag optional must not make its typo check optional.
-        let err = validate_serve_args(&parse(&["--mtp-gate", "always"])).unwrap_err();
-        assert!(err.contains("--mtp-gate"), "{err}");
-        assert!(err.contains("auto, force"), "names the valid values: {err}");
-    }
-
-    #[test]
-    fn require_auth_needs_a_token() {
-        assert!(validate_serve_args(&parse(&["--require-auth"])).is_err());
-        assert!(validate_serve_args(&parse(&["--require-auth", "--auth-token", "sk-x"])).is_ok());
-    }
-
-    #[test]
-    fn num_drafts_needs_speculative() {
-        assert!(validate_serve_args(&parse(&["--num-drafts", "2"])).is_err());
-        assert!(validate_serve_args(&parse(&["--num-drafts", "2", "--speculative"])).is_ok());
-    }
-
-    #[test]
-    fn rank_must_be_below_world_size() {
-        assert!(validate_serve_args(&parse(&["--rank", "2", "--world-size", "2"])).is_err());
-        assert!(validate_serve_args(&parse(&["--rank", "1", "--world-size", "2"])).is_ok());
-    }
-
-    #[test]
-    fn ep_size_cannot_exceed_world_size() {
-        assert!(validate_serve_args(&parse(&["--ep-size", "2"])).is_err());
-        assert!(validate_serve_args(&parse(&["--ep-size", "2", "--world-size", "2"])).is_ok());
-    }
-
-    #[test]
-    fn disable_thinking_conflicts_with_budget() {
-        assert!(
-            validate_serve_args(&parse(&[
-                "--disable-thinking",
-                "--max-thinking-budget",
-                "2048"
-            ]))
-            .is_err()
-        );
-    }
-
-    #[test]
-    fn flagship_recipe_is_accepted() {
-        // The canonical 35B flagship serve recipe (PR #278) passes
-        // `--kv-cache-dtype bf16 --kv-high-precision-layers auto` together —
-        // redundant but valid. The validator must NOT reject it.
-        assert!(
-            validate_serve_args(&parse(&[
-                "--kv-cache-dtype",
-                "bf16",
-                "--lm-head-dtype",
-                "nvfp4",
-                "--kv-high-precision-layers",
-                "auto",
-                "--scheduling-policy",
-                "slai",
-                "--speculative",
-                "--num-drafts",
-                "1",
-                "--mtp-quantization",
-                "bf16",
-                "--enable-prefix-caching",
-            ]))
-            .is_ok()
-        );
-    }
-
-    #[test]
-    fn enum_typos_are_rejected() {
-        let err = validate_serve_args(&parse(&["--scheduling-policy", "fifoo"])).unwrap_err();
-        assert!(err.contains("--scheduling-policy"));
-        assert!(err.contains("fifo, slai"));
-    }
-
-    #[test]
-    fn multiple_violations_all_reported() {
-        let err = validate_serve_args(&parse(&[
-            "--require-auth",
-            "--num-drafts",
-            "3",
-            "--rank",
-            "5",
-            "--world-size",
-            "2",
-        ]))
-        .unwrap_err();
-        assert!(err.contains("[1]"));
-        assert!(err.contains("[2]"));
-        assert!(err.contains("[3]"));
-    }
-
-    #[test]
-    fn gpu_mem_util_range_enforced() {
-        assert!(validate_serve_args(&parse(&["--gpu-memory-utilization", "1.5"])).is_err());
-        assert!(validate_serve_args(&parse(&["--gpu-memory-utilization", "0.0"])).is_err());
-        assert!(validate_serve_args(&parse(&["--gpu-memory-utilization", "0.9"])).is_ok());
-    }
-}
+#[path = "validate_tests.rs"]
+mod tests;
